Replace PointsAdminPanel with LessonResultsPanel
Removed PointsAdminPanel and introduced LessonResultsPanel for lesson results administration. Updated routing and lazy imports to use the new panel. Adjusted related styles and minor UI logic. Also fixed a z-index issue in LeadForm.styled.js and improved document title formatting in Stream.jsx.
