+++ src/components/LeadForm/LeadForm.styled.js
@@ -264,7 +264,6 @@ export const FormBtn = styled.button`
 
 export const FormBtnText = styled.span`
   position: absolute;
-  z-index: 5;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
